requirements.txt
aiogram
fastapi==0.111.0
uvicorn==0.30.1
weasyprint
jinja2
python-dotenv